chore: refactor pipeline forwarders (#2925)
