+++ rust/Makefile
@@ -15,7 +15,11 @@ test:
 
 .PHONY: all-tests
 all-tests:
-	cargo test --all-features
+	cargo test --features nats-tests,pulsar-tests,sqs-tests,kafka-tests --workspace --all
+	cargo test --features panic-tests test_map_stream_with_panic
+	#cargo test --features panic-tests test_batch_map_with_panic
+	cargo test --features panic-tests test_map_stream_panic
+	cargo test --features panic-tests test_transform_stream_with_panic
 
 .PHONY: clean
 clean:
