+++ rust/numaflow-core/Cargo.toml
@@ -8,7 +8,8 @@ nats-tests = []
 pulsar-tests = []
 sqs-tests = []
 kafka-tests = []
-all-tests = ["nats-tests", "pulsar-tests", "sqs-tests", "kafka-tests"]
+panic-tests = []
+all-tests = ["nats-tests", "pulsar-tests", "sqs-tests", "kafka-tests", "panic-tests"]
 
 [lints]
 workspace = true
@@ -64,7 +65,7 @@ async-trait = "0.1.88"
 
 [dev-dependencies]
 tempfile = "3.11.0"
-numaflow = { git = "https://github.com/numaproj/numaflow-rs.git", rev = "9a12a1d832d4fece830452044bd7706dc8b2a381" }
+numaflow = "0.3.0"
 pulsar = { version = "6.3.0", default-features = false, features = ["tokio-rustls-runtime"] }
 hyper = "1.6.0"
 hyper-rustls = "0.27.5"
