+++ rust/numaflow-core/src/reduce/reducer/unaligned/user_defined/accumulator.rs
@@ -243,7 +243,7 @@ mod tests {
                     let keys = request.keys.clone();
                     let updated_key = format!("{}_{}", keys.join(":"), current_count);
 
-                    let mut message = accumulator::Message::from_datum(request);
+                    let mut message = accumulator::Message::from_accumulator_request(request);
                     message = message.with_keys(vec![updated_key]);
                     message = message.with_value(format!("count_{}", current_count).into_bytes());
 
@@ -287,7 +287,7 @@ mod tests {
             value: "value1".into(),
             offset: Offset::String(StringOffset::new("0".to_string(), 0)),
             event_time: Utc.with_ymd_and_hms(2023, 1, 1, 0, 0, 10).unwrap(),
-            watermark: None,
+            watermark: Some(Utc.with_ymd_and_hms(2023, 1, 1, 0, 0, 0).unwrap()),
             id: MessageID {
                 vertex_name: "vertex_name".to_string().into(),
                 offset: "0".to_string().into(),
@@ -303,7 +303,7 @@ mod tests {
             value: "value2".into(),
             offset: Offset::String(StringOffset::new("1".to_string(), 1)),
             event_time: Utc.with_ymd_and_hms(2023, 1, 1, 0, 0, 20).unwrap(),
-            watermark: None,
+            watermark: Some(Utc.with_ymd_and_hms(2023, 1, 1, 0, 0, 0).unwrap()),
             id: MessageID {
                 vertex_name: "vertex_name".to_string().into(),
                 offset: "1".to_string().into(),
@@ -319,7 +319,7 @@ mod tests {
             value: "value3".into(),
             offset: Offset::String(StringOffset::new("2".to_string(), 2)),
             event_time: Utc.with_ymd_and_hms(2023, 1, 1, 0, 0, 30).unwrap(),
-            watermark: None,
+            watermark: Some(Utc.with_ymd_and_hms(2023, 1, 1, 0, 0, 0).unwrap()),
             id: MessageID {
                 vertex_name: "vertex_name".to_string().into(),
                 offset: "2".to_string().into(),
@@ -453,7 +453,7 @@ mod tests {
                 value: "valueA1".into(),
                 offset: Offset::String(StringOffset::new("0".to_string(), 0)),
                 event_time: Utc.with_ymd_and_hms(2023, 1, 1, 0, 0, 10).unwrap(),
-                watermark: None,
+                watermark: Some(Utc.with_ymd_and_hms(2023, 1, 1, 0, 0, 0).unwrap()),
                 id: MessageID {
                     vertex_name: "vertex_name".to_string().into(),
                     offset: "0".to_string().into(),
@@ -468,7 +468,7 @@ mod tests {
                 value: "valueB1".into(),
                 offset: Offset::String(StringOffset::new("1".to_string(), 1)),
                 event_time: Utc.with_ymd_and_hms(2023, 1, 1, 0, 0, 20).unwrap(),
-                watermark: None,
+                watermark: Some(Utc.with_ymd_and_hms(2023, 1, 1, 0, 0, 0).unwrap()),
                 id: MessageID {
                     vertex_name: "vertex_name".to_string().into(),
                     offset: "1".to_string().into(),
@@ -483,7 +483,7 @@ mod tests {
                 value: "valueA2".into(),
                 offset: Offset::String(StringOffset::new("2".to_string(), 2)),
                 event_time: Utc.with_ymd_and_hms(2023, 1, 1, 0, 0, 30).unwrap(),
-                watermark: None,
+                watermark: Some(Utc.with_ymd_and_hms(2023, 1, 1, 0, 0, 0).unwrap()),
                 id: MessageID {
                     vertex_name: "vertex_name".to_string().into(),
                     offset: "2".to_string().into(),
@@ -498,7 +498,7 @@ mod tests {
                 value: "valueA3".into(),
                 offset: Offset::String(StringOffset::new("3".to_string(), 3)),
                 event_time: Utc.with_ymd_and_hms(2023, 1, 1, 0, 0, 40).unwrap(),
-                watermark: None,
+                watermark: Some(Utc.with_ymd_and_hms(2023, 1, 1, 0, 0, 0).unwrap()),
                 id: MessageID {
                     vertex_name: "vertex_name".to_string().into(),
                     offset: "3".to_string().into(),
@@ -513,7 +513,7 @@ mod tests {
                 value: "valueB2".into(),
                 offset: Offset::String(StringOffset::new("4".to_string(), 4)),
                 event_time: Utc.with_ymd_and_hms(2023, 1, 1, 0, 0, 50).unwrap(),
-                watermark: None,
+                watermark: Some(Utc.with_ymd_and_hms(2023, 1, 1, 0, 0, 0).unwrap()),
                 id: MessageID {
                     vertex_name: "vertex_name".to_string().into(),
                     offset: "4".to_string().into(),
@@ -528,7 +528,7 @@ mod tests {
                 value: "valueB3".into(),
                 offset: Offset::String(StringOffset::new("5".to_string(), 5)),
                 event_time: Utc.with_ymd_and_hms(2023, 1, 1, 0, 1, 0).unwrap(),
-                watermark: None,
+                watermark: Some(Utc.with_ymd_and_hms(2023, 1, 1, 0, 0, 0).unwrap()),
                 id: MessageID {
                     vertex_name: "vertex_name".to_string().into(),
                     offset: "5".to_string().into(),
