+++ rust/numaflow-core/src/transformer.rs
@@ -517,6 +517,7 @@ mod tests {
         }
     }
 
+    #[cfg(feature = "panic-tests")]
     #[tokio::test]
     async fn test_transform_stream_with_panic() -> Result<()> {
         let tmp_dir = TempDir::new().unwrap();
